[FLINK-10922] Refactor the placement of the Flink Kafka connector end to end test module
