+++ flink-end-to-end-tests/pom.xml
@@ -61,6 +61,10 @@ under the License.
 		<module>flink-metrics-availability-test</module>
 		<module>flink-metrics-reporter-prometheus-test</module>
 		<module>flink-heavy-deployment-stress-test</module>
+		<module>flink-streaming-kafka-test-base</module>
+		<module>flink-streaming-kafka-test</module>
+		<module>flink-streaming-kafka011-test</module>
+		<module>flink-streaming-kafka010-test</module>
 	</modules>
 
 	<build>
